Cache property rename - PartitionKey to Name
This was initially used as a way to distinguish between the long-term cache and the short-term temporary cache when naming the cache files, however the short-term cache is always a memory-only cache so the need for a 'partition' no longer exists.
Now just a useful way of naming the instance of the memory cache.